requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
jinja2==3.1.4
python-multipart==0.0.17
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
starlette==0.41.3
httpx==0.28.0
python-dotenv==1.0.1
groq==0.13.0
cryptography==43.0.3
aiofiles==24.1.0
schedule==1.2.2
email-validator==2.2.0
